+++ features/settings/settingsPanel.js
@@ -506,7 +506,7 @@
       )
       + '</div>';
   }
-  var APP_VERSION = '0.1.0128';
+  var APP_VERSION = '0.1.0129';
 
 
   function healthSummaryPill(){